separate out user file, use reply.h
